+++ src/ports/MilitaryAirport.java
@@ -2,6 +2,7 @@ package ports;
 
 import enumerates.typesOfArms;
 import interfaces.planeFactory;
+import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.IntegerProperty;
 import javafx.util.Pair;
 import vehicles.Aircraft;
@@ -10,12 +11,12 @@ import other.TravelRoute;
 
 public class MilitaryAirport extends Airport implements planeFactory{
 
-    public MilitaryAirport(String name, Pair<IntegerProperty, IntegerProperty> coordinates) {
+    public MilitaryAirport(String name, Pair<DoubleProperty, DoubleProperty> coordinates) {
         super(name, coordinates);
     }
 
     @Override
-    public Aircraft createAircraft(Pair<IntegerProperty, IntegerProperty> coordinates, int id, IntegerProperty maximumAmountOfFuel,
+    public Aircraft createAircraft(Pair<DoubleProperty, DoubleProperty> coordinates, int id, IntegerProperty maximumAmountOfFuel,
                                    IntegerProperty currentAmountOfFuel, int amountOfStaff, Airport lastVisitedAirport, Airport nextAirport,
                                    TravelRoute travelRoute, typesOfArms typeOfArms)
     {
